Ascendent powernode creation implemented through include_block atoms
A new powernode can contains already existing powernodes, while these
are entirely contained.
This behavior is achieve by construct and maintain a tree of inclusion
of blocks with the include_block atom. The topblock atom is introduced
for allow a newconcept to find its smaller container, that will be
modified in the inclusion tree: it will have a new children, the
newconcept, and can lose others blocks that are contained by the
newconcept.