New option: can use non-threaded method solving
This is something like a big update : the powergrasp modules don't
handle models anymore, but only lists of atoms.
The commons module allow client code to choose the method of solving
between threaded and no-threaded. The second is slower because of a
weird behavior of the Control.solve(2) method provided by python-gringo.