fix bug#6
Now the node is not placed in a new block if it is alone in its concept
set. (star/3)
A new test case for test the star case is created, and show that nodes
that are use only in star as single nodes are not present in the NODE
list in bubble output.