first implementation, upper bound seems to be inadequate in many cases
This implementation compute the minimal upper bound in the scorebound
program, for each node.
When newedgecover is computed in the findbestconcept program, the score
must be smaller than minimal upper bound associated to the minobj of the
newconcept.
This doesn't work because the minupperbound doesn't exists for all minobj,
and this case is not handled by newedgecover definition.