naming: define exactly diff between score and edge cover
edge cover equals to number of covered edge, and must be maximized.
for allow the solver to use a minimization instead of a maximization
(which is more complicated and possibly non-optimized), the score is
defined as a to minimize value, depending of the maximal edge cover and
the new concept edge cover.