replace block atoms and related by a dynamic test of overlapping
This is the main modification introduced by the optimization
enhancement#15. It's currently not interesting for runtime optimization,
but it can be widely improved through some python optimizations.
Note that the newconcept is now named concept, and this concept is first
modified for being valid. Then, it score will be computed. The
maximal constraint of generated concepts can be applied for avoid a two
big quantity of initially generated concepts.