rewrite: stars + powernode reusing + inclusions
Based on a code review and many existing bugs, the code is heavily
rewrited for take count of stars motifs that appears after the concept
reduction. Moreover, the existing powernodes are reused if possible, and
the inclusions get some bugfixes. Currently the compression is longuer,
and generate in some non-understandable case a bad compression and
errors (especially *cccoli*).