Fix a bug about the FinalEdgeCountGenerated
Signal was sent by compression, and was intercepted by statistics
DataExtractor object, which take the data given by the signal (that
was equals to the number of remain edge in the cc, not the whole graph)
and put it in global stats.
Result was a number of remaining edges in the final stats that fall from
normal value to 1 or 0 (number of edges remaining in the cc),
and then rise to a normal value.