First implementation of equivalent node detection and usage
The detection is fully functional: yields weight/2 and equiv/2 atoms,
and hide nodes equivalent to a smaller one through the hidden/1 atom.
Then, extraction only yields nodes interesting for compression.

However, some cases shows that equivalences, in cases of cliques
finding, can be invalidated. Updates of the lattice are needed.

Moreover, the usage of weights for determine the score of a clique is
not properly implemented. With this optimization, no concept is detected.