tests: clique_reduction: fix test case
This test case shows the problem of cliques in case of node reduction.
The clique is formed before any other concept, using the node a, that
hides the node b. The retained solution is to use b instead of a, and
reduce the weight of a by 1. Atom equiv(a,b) must also be invalidated.