ASP:extract: edge/2 defines in which set belongs each node
this is the important part of the optimization : the graph reduction
algorithm defined in powergrasp/graph_reduction.py file create two sets
of nodes, where nodes of the same set are always together in the
biclique generation. This is basically a pre-generation of the bicliques
sets. For mark that, edge(X,Y) defines X as a node in set 1, and Y a
node in set 2.