test case: remove reserved keywords from the test case
this is not the first time a **very strange** bug appears,
du to the usage of a node name like <k> or <cc>, which are converted to
constant during the compression, leading to nodes of id equals to the
current step or cc.
This time, this bug is properly treated: a critical error is raised if
any node is named with a reserved keyword.

Hope it will never happen again.