solving: constants are written in file instead of passed by cli
this is needed in order to avoid problems with cli handling. In fact,
a node with a very weird name (like '; rm -rf /* ; echo "') could happen
to be destructive for host.
(or the less dangerous case of a node having spaces and other bash
 related stuff in its name,
 like "1,3-dimethyl-2-[2-oxopropyl thio]imidazolium chloride")

Consequently, now the constants are passed as additional program text
to pyasp, written in asp using #const meta-instruction.
Note that the main downside of this technic is that, if the constant is
already specified in the file using the #const, there will be a clash
because of constant redefinition.
This do not happen (currently) in the powergrasp codebase, but i bet it
will happen one day, and that i will lost precious time because of that.

Sorry, me.